Designate a set of mapped items.
New shape for mapped results. It is an entry with the sought key and an
array of items that matched the possibly partial key. Thus, zero, one,
or more. The order of items is always forward. We perform our forward
test on each entry. Disucssions of this change can be found in the
commit log entries prior to the commits Homogenize release
v2.0.0-alpha.10 and Skip release v0.2.7.

https://github.com/bigeasy/homogenize/tree/v2.0.0-alpha.10
https://github.com/bigeasy/skip/releases/tag/v0.2.7

Closes #122.